requirements.txt
ijson
pandas
psycopg2
python-dotenv
requests
sqlalchemy
thefuzz
tqdm
